+++ library/src/main/java/com/avast/android/dialogs/core/BaseDialogFragment.java
@@ -16,6 +16,10 @@
 
 package com.avast.android.dialogs.core;
 
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
+
 import android.app.Dialog;
 import android.content.Context;
 import android.content.DialogInterface;
@@ -23,6 +27,7 @@ import android.graphics.Typeface;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.v4.app.DialogFragment;
+import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentManager;
 import android.support.v4.app.FragmentTransaction;
 import android.text.TextUtils;
@@ -32,14 +37,17 @@ import android.view.ViewGroup;
 import android.widget.*;
 
 import com.avast.android.dialogs.R;
+import com.avast.android.dialogs.iface.ISimpleDialogCancelListener;
 import com.avast.android.dialogs.util.TypefaceHelper;
 
 /**
- * Base dialog fragment for all your dialogs, stylable and same design on Android 2.2+.
+ * Base dialog fragment for all your dialogs, styleable and same design on Android 2.2+.
  *
  * @author David Vávra (david@inmite.eu)
  */
 public abstract class BaseDialogFragment extends DialogFragment implements DialogInterface.OnShowListener {
+
+    protected int mRequestCode;
 
     @NonNull
     @Override
@@ -61,6 +69,27 @@ public abstract class BaseDialogFragment extends DialogFragment implements Dialo
         return build(builder).create();
     }
 
+    @Override
+    public void onActivityCreated(Bundle savedInstanceState) {
+        super.onActivityCreated(savedInstanceState);
+        final Fragment targetFragment = getTargetFragment();
+        if (targetFragment != null) {
+            mRequestCode = getTargetRequestCode();
+        } else {
+            Bundle args = getArguments();
+            if (args != null) {
+                mRequestCode = args.getInt(BaseDialogBuilder.ARG_REQUEST_CODE, 0);
+            }
+        }
+    }
+
+    /**
+     * Key method for using {@link com.avast.android.dialogs.core.BaseDialogFragment}.
+     * Customized dialogs need to be set up via provided builder.
+     *
+     * @param initialBuilder Provided builder for setting up customized dialog
+     * @return Updated builder
+     */
     protected abstract Builder build(Builder initialBuilder);
 
     @Override
@@ -81,10 +110,60 @@ public abstract class BaseDialogFragment extends DialogFragment implements Dialo
     @Override
     public void onShow(DialogInterface dialog) {
         if (getView() != null) {
-            ScrollView vScrollView = (ScrollView)getView().findViewById(R.id.sdl_scrollview);
-            boolean scrollable = isScrollable(vScrollView);
+            ScrollView vMessageScrollView = (ScrollView)getView().findViewById(R.id.sdl_message_scrollview);
+            ListView vListView = (ListView)getView().findViewById(R.id.sdl_list);
+            FrameLayout vCustomViewNoScrollView = (FrameLayout)getView().findViewById(R.id.sdl_custom);
+            boolean customViewNoScrollViewScrollable = false;
+            if (vCustomViewNoScrollView.getChildCount() > 0) {
+                View firstChild = vCustomViewNoScrollView.getChildAt(0);
+                if (firstChild instanceof ViewGroup) {
+                    customViewNoScrollViewScrollable = isScrollable((ViewGroup)firstChild);
+                }
+            }
+            boolean listViewScrollable = isScrollable(vListView);
+            boolean messageScrollable = isScrollable(vMessageScrollView);
+            boolean scrollable = listViewScrollable || messageScrollable || customViewNoScrollViewScrollable;
             modifyButtonsBasedOnScrollableContent(scrollable);
         }
+    }
+
+    @Override
+    public void onCancel(DialogInterface dialog) {
+        super.onCancel(dialog);
+        for (ISimpleDialogCancelListener listener : getCancelListeners()) {
+            listener.onCancelled(mRequestCode);
+        }
+    }
+
+    /**
+     * Get dialog cancel listeners.
+     * There might be more than one cancel listener.
+     *
+     * @return Dialog cancel listeners
+     * @since 2.1.0
+     */
+    protected List<ISimpleDialogCancelListener> getCancelListeners() {
+        return getDialogListeners(ISimpleDialogCancelListener.class);
+    }
+
+    /**
+     * Utility method for acquiring all listeners of some type for current instance of DialogFragment
+     *
+     * @param listenerInterface Interface of the desired listeners
+     * @return Unmodifiable list of listeners
+     * @since 2.1.0
+     */
+    @SuppressWarnings("unchecked")
+    protected <T> List<T> getDialogListeners(Class<T> listenerInterface) {
+        final Fragment targetFragment = getTargetFragment();
+        List<T> listeners = new ArrayList<T>(2);
+        if (targetFragment != null && listenerInterface.isAssignableFrom(targetFragment.getClass())) {
+            listeners.add((T)targetFragment);
+        }
+        if (getActivity() != null && listenerInterface.isAssignableFrom(getActivity().getClass())) {
+            listeners.add((T)getActivity());
+        }
+        return Collections.unmodifiableList(listeners);
     }
 
     /**
@@ -111,12 +190,12 @@ public abstract class BaseDialogFragment extends DialogFragment implements Dialo
         }
     }
 
-    boolean isScrollable(ScrollView scrollView) {
-        if (scrollView.getChildCount() == 0) {
-            return false;
+    private boolean isScrollable(ViewGroup listView) {
+        int totalHeight = 0;
+        for (int i = 0; i < listView.getChildCount(); i++) {
+            totalHeight += listView.getChildAt(i).getMeasuredHeight();
         }
-        final int childHeight = scrollView.getChildAt(0).getMeasuredHeight();
-        return scrollView.getMeasuredHeight() < childHeight;
+        return listView.getMeasuredHeight() < totalHeight;
     }
 
     /**
@@ -152,8 +231,11 @@ public abstract class BaseDialogFragment extends DialogFragment implements Dialo
 
         private int mListCheckedItemIdx;
 
-        private AdapterView.OnItemClickListener mOnItemClickListener;
+        private int mChoiceMode;
 
+        private int[] mListCheckedItemMultipleIds;
+
+        private AdapterView.OnItemClickListener mOnItemClickListener;
 
 
         public Builder(Context context, LayoutInflater inflater, ViewGroup container) {
@@ -222,6 +304,16 @@ public abstract class BaseDialogFragment extends DialogFragment implements Dialo
             return this;
         }
 
+
+        public Builder setItems(ListAdapter listAdapter, int[] checkedItemIds, int choiceMode, final AdapterView.OnItemClickListener listener) {
+            mListAdapter = listAdapter;
+            mListCheckedItemMultipleIds = checkedItemIds;
+            mOnItemClickListener = listener;
+            mChoiceMode = choiceMode;
+            mListCheckedItemIdx = -1;
+            return this;
+        }
+
         /**
          * Set list
          *
@@ -232,6 +324,7 @@ public abstract class BaseDialogFragment extends DialogFragment implements Dialo
             mListAdapter = listAdapter;
             mOnItemClickListener = listener;
             mListCheckedItemIdx = checkedItemIdx;
+            mChoiceMode = AbsListView.CHOICE_MODE_NONE;
             return this;
         }
 
@@ -272,6 +365,12 @@ public abstract class BaseDialogFragment extends DialogFragment implements Dialo
                 vList.setOnItemClickListener(mOnItemClickListener);
                 if (mListCheckedItemIdx != -1) {
                     vList.setSelection(mListCheckedItemIdx);
+                }
+                if (mListCheckedItemMultipleIds != null) {
+                    vList.setChoiceMode(mChoiceMode);
+                    for (int i : mListCheckedItemMultipleIds) {
+                        vList.setItemChecked(i, true);
+                    }
                 }
             }